fix:resolve product details redirect and cart count disappearing on page refresh
